Refactor Converter class into separate modules
The utils/ directory was created to allow the separation of
functionality but everything ended up in a single class for ease. Before
the repo grows too large let's split up as planned.
